+++ R/src-sql.r
@@ -72,7 +72,7 @@ src_tbls.src_sql <- function(x, ...) {
 #' @export
 format.src_sql <- function(x, ...) {
   paste0(
-    "src:  ", src_desc(x), "\n",
+    "src:  ", db_desc(x$con), "\n",
     wrap("tbls: ", paste0(sort(src_tbls(x)), collapse = ", "))
   )
 }